+++ src/quickSearcherProvider.ts
